requirements.txt
langfuse
tiktoken
openai
python-dotenv